ai-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
tavily-python==0.3.0
langchain==0.1.20
langchain-community==0.0.38
mysql-connector-python==8.2.0
python-multipart==0.0.6
requests==2.31.0
ollama==0.1.7